url/requirements.txt
blinker
click
colorama
contourpy
cycler
Flask
Flask-Cors
fonttools
gunicorn
itsdangerous
Jinja2
joblib==1.0.0
kiwisolver
MarkupSafe
matplotlib
numpy
packaging
pandas
Pillow
pyparsing
python-dateutil
python-whois
pytz
scikit-learn
scipy
six
threadpoolctl
tzdata
Werkzeug
xgboost==1.6.2
